ModHistory: Add a wrapper for DiffRawView
To make the Diff usable, we need to append at least a combo box on top
of it.
There we need to select to which parent we want to diff (Or to 'all'
parents). And if we got it we can also add to that Combo box:
- HEAD
- Index
- SHA1 (Needs an additional QLineEdit)

This patch adds the wrapper