GitWrap: Treat Config as a BasicObject
LibGit2-Devs tend to produce beasts - Or maybe that's just due to them
living in their little C-Based world, where they bravely get and eat
their pills every morning and don't have to bother with real world
concepts like object oriented programming...

Config sometimes is depending on a Repository and sometimes not. This
patch doesn't change GitWrap's semantics. Should add a GitPtr to the
Repository and only fill (and reference it) if the Config is really
assigned to a Repository. TreeEntry is a similar beast of it's own:
Depending on where it comes from, we either MUST free it or are not
even ALLOWED to free it.