Revert faster MOC-handling.
The reason for reverting the faster MOC handling is that it needs to
declare a file as GENERATED, which actually means "is generated by the
build". But in reality this file is generated during the cmake phase.

This, however, breaks the "clean" target. (make clean removes the file,
thinking it can be regenerated while it actually can't. A new cmake run
is required to build it. But cmake will update all those moc files,
causing a rebuild of all mocs)