Fix an error in History view
History view would only reset it's model when switching to an invalid
repository. This is wrong (as for a valid repository the model will soon
be rebuilt after a repo switch).

It led me to think that contextual switches are a requirement for my
topic/uncrappify1 changes to work. As it seems, this already makes the
uncrappify changes look pretty good.