Much less queued events in History View
Up to now, every row that was not yet populated caused approximately 20
events to be queued into the event queue and thus was followed by 20
times checking if it was already populated.

This change uses an addition bit in the entry to store whether we're
enqueued for populating. Now the model's data() method will do only one
request for populating the entry. In theory this renders the previous
check in HistoryModel::ensurePopulated() useless. It is still left
there, just for paranoia.