RM: Refactor the refactoring branch :-[
This effectively introduces Private-Classes for all RepoMan objects.
This pile of shit is necessary, to create central method similar to
preTerminate(), but for the opposite purpose: postCreation().

It has to be called when all of the outer-most object is intialized but
must be called virtually from the base. This is only possible if the
object is separated into a data container and the public interface.

Thus, we're having the private-class separation now in heaven, gitwrap
and libgitcore. When I start doing something that big next time, I
_should_ really start with that pattern right from the start, shouldn't
I?