Adjust RCC-Macro to workaround CMake 2.8.12 changes
Actually, this uses the unmodified CMake Macro on 2.8.12 and uses
our improved one with prior version of CMake. This is not ideal, but
changing established macros isn't either. I'd bet that if you'd ask
those guys, they wouldn't bother a second to argue that these macros
are meant to be private...