Link gtest to pthreads on linux/macosx
Also: don't link static libraries privately to their dependencies. It
must be public, because we have to name them actually just so that they
can be used when linking against the static library (This is exactly
public linking).

This change is required as the gtest code expects to pthreads to be
available unless we tell it otherwise. Strangely, this doesn't fix the
broken builds in my existing trees - and new trees seem unaffected
	(Which is another strange thing).