RM: Don't recreate all submodules all the time
We are stripping the trailing slash from the path that libgit2 gave us
when we create the RepoMan object. When we refresh the submodule, we
don't do that, but still compare the paths for equality. Thus, on every
refresh, we recreate every submodule.

Actually, this was a bug in untested old code. Damn it.