removed brows button for the "from" field
Cloning a local repository is not usual. It only makes sense when doing a bare clone. But even then you'll be always better off by creating the bare repository on the remote end, then add a remote locally and push afterwards.

It's still possible to do though. You just have to type in the full local path.